+++ CHANGELOG.md
@@ -3,11 +3,15 @@ CHANGELOG
 
 unreleased
 ----------
-- Use version 3.20.0 of braintree-web
-- Update browser-detection to v1.5.0
+- Add `cardholderName` option to card configuration
+
+1.5.0
+-----
+- Use version 3.20.1 of braintree-web
+- Update browser-detection to v1.6.0
 - Add `aria-label` attribute to payment options
 - Update checkout.js to v4.0.95
-- Add `cardholderName` option to card configuration
+- Add `clearSelectedPaymentMethod` to remove selected payment method
 
 1.4.0
 ------